requirements.txt
absl-py==1.4.0
alembic==1.12.0
antlr4-python3-runtime==4.9.3
appdirs==1.4.4
astunparse==1.6.3
attrs==23.1.0
audioread==3.0.0
autopage==0.5.1
beautifulsoup4==4.12.2
cachetools==5.3.1
certifi==2023.5.7
cffi @ file:///croot/cffi_1670423208954/work
charset-normalizer==3.1.0
click==8.1.3
cliff==4.3.0
cmaes==0.10.0
cmake==3.26.4
cmd2==2.4.3
colorlog==6.7.0
contourpy==1.1.0
cycler==0.11.0
Cython==0.29.35
cytoolz==0.12.1
dataclasses==0.6
decorator==5.1.1
einops==0.6.1
fastdtw==0.3.4
filelock==3.12.2
flatbuffers==23.5.26
fonttools==4.40.0
fsspec==2023.6.0
future @ file:///croot/future_1677599870788/work
gast==0.4.0
gdown==4.7.1
gmpy2 @ file:///tmp/build/80754af9/gmpy2_1645455533097/work
google-auth==2.20.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.56.0
h5py==3.9.0
huggingface-hub==0.15.1
hydra-core==1.3.2
hydra-optuna-sweeper==1.2.0
HyperPyYAML==1.2.1
idna==3.4
importlib-metadata==6.8.0
inflect==6.0.4
intervaltree==3.1.0
Jinja2 @ file:///croot/jinja2_1666908132255/work
jiwer==3.0.3
joblib==1.2.0
keras==2.13.1
kiwisolver==1.4.4
lazy_loader==0.2
lhotse==1.15.0
libclang==16.0.0
librosa==0.10.0.post2
lilcom==1.6
lit==16.0.6
llvmlite==0.40.1
Mako==1.2.4
Markdown==3.4.3
MarkupSafe==2.1.3
matplotlib==3.7.1
mkl-fft==1.3.6
mkl-random @ file:///work/mkl/mkl_random_1682950433854/work
mkl-service==2.4.0
ml-dtypes==0.2.0
mpmath==1.3.0
msgpack==1.0.5
networkx==3.1
ninja==1.11.1
numba==0.57.1
numpy==1.24.3
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
omegaconf==2.3.0
opt-einsum==3.3.0
optuna==2.10.1
packaging==23.1
pandas==2.0.2
pbr==5.11.1
Pillow==10.0.0
pooch==1.6.0
prettytable==3.8.0
protobuf==4.23.3
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser @ file:///tmp/build/80754af9/pycparser_1636541352034/work
pydantic==1.10.9
pydub==0.25.1
pyparsing==3.1.0
pyperclip==1.8.2
PySocks==1.7.1
pysptk==0.2.0
python-dateutil==2.8.2
pytz==2023.3
pyworld==0.3.4
PyYAML @ file:///croot/pyyaml_1670514731622/work
rapidfuzz==3.2.0
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
ruamel.yaml==0.17.28
ruamel.yaml.clib==0.2.7
scikit-learn==1.2.2
scipy==1.10.1
sentencepiece==0.1.99
six==1.16.0
sortedcontainers==2.4.0
soundfile==0.12.1
soupsieve==2.4.1
soxr==0.3.5
speechbrain==0.5.14
sphfile==1.0.3
SQLAlchemy==2.0.20
stevedore==5.1.0
sympy==1.12
tabulate==0.9.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-io-gcs-filesystem==0.32.0
termcolor==2.3.0
threadpoolctl==3.1.0
toolz==0.12.0
torch==2.0.1
torch-tb-profiler==0.4.1
torchaudio==2.0.2
torchvision==0.15.2
tqdm==4.65.0
triton==2.0.0
typing_extensions==4.5.0
tzdata==2023.3
Unidecode==1.3.6
urllib3==1.26.16
wcwidth==0.2.6
Werkzeug==2.3.6
wrapt==1.15.0
zipp==3.16.2
